Update ToltecSignalFits.py
Changes to incorporate list of ToltecSignalFits in the class ToltecListFits and being able to compare observations. Some changes in the fitGaussian method, since it was giving me errors and bad fits. The beam is needed to perform point source photometry, although it should be changed in the future to use the units of MJy/Sr